Started restructuring the code to make it work with multiple images at a time.
I have now begun implementing the functionality of working with folder of images, instead of just single images at a time.

To make it simple i have some assumptions about the structure of the folders, and then iterate over all the files in the folder given as an argument in the terminal.